### 2017-09-24 #414 @whedon
[PRE REVIEW]: Augmentor: An Image Augmentation Library for Machine Learning
**Submitting author:** @mdbloice (<a href="http://orcid.org/0000-0002-2468-4086">Marcus Bloice</a>)
**Repository:** <a href="https://github.com/mdbloice/Augmentor" target ="_blank">https://github.com/mdbloice/Augmentor</a>
**Version:** v0.1.5
