### 2017-10-17 #414 @arfon
@sealhuang - would you be willing to review this submission for JOSS?
