### 2017-10-22 #432 @sealhuang
I have completed the review, @arfon and @mdbloice.

The Augmentor package provides a convenient and flexible approach to generate new image samples via various operations. The package is particularly useful for generating batch of sample while training a deep learning model.

Overall I think this is a submission suited for the JOSS, and it meets the requirements for acceptance with a few exceptions. I've detailed the items below

- [ ] Version: The latest stable version is 0.1.18, not 0.1.5

- [ ] Authors: There are 3 authors listed in the paper.md (minor point: no affiliation information for the 2nd author), however, the commit logs show some other people also contribute codes to the project. If they are not listed as authors, I think they should be referenced/acknowledged somewhere in the docs as contributors explicitly.

- [ ] Automated tests: Although the test code is included in the repo, however, the document about how to use the test code is still lacking (i.e. how to use pytest to verify the software). In addition, the tests code is not installed via pip installation approach. I think it would be better that moving the `tests` directory into `Augmentor`, and installing the core code and tests code together. And a test API (something like Augmentor.test()) may be more user-friendly.

This is my first review with JOSS, so please tell me if I am doing something incorrectly. @arfon
