### 2017-10-24 #432 @arfon
@sealhuang - many thanks for the review and your thoughtful comments ✨ . @mdbloice - over to you for your responses.
