### 2017-10-31 #432 @mdbloice
Hi @sealhuang @arfon many thanks for the review and organisation!

I will try to address each point here:

- Version: Since submitting to JOSS I made a few bug fixes, which were quite important for a few features of the Augmentor. As they were quite crucial I wanted the changes to be on the PyPI, hence the version number discrepancy. These changes were quite minor, and feature-wise there is very little difference between v0.1.5 and v0.1.8.
- Authors: Yes, author 2 does not have an academic affiliation so this was purposely left as it is. As for the contributors, is not not enough that they are listed in the Contributors tab on the repo homepage? I would guess this list will change fairly often, so I think it would be wasted effort to maintain my own list somewhere such as in the README would it not?
- Automated tests: I would prefer not to include the tests in the pip installation, however, I have now added a section to the README which describes how one can perform the tests locally by cloning the repository. I hope that this compromise is OK with you @sealhuang ?

If anything requires clarification please let me know and I will respond as quickly as I can, and may I also apologise for the lateness of my reply, I have been inundated at work of late :)

Again many thanks to both for the review and efforts,

Marcus.
