### 2017-10-31 #432 @arfon
> Authors: Yes, author 2 does not have an academic affiliation so this was purposely left as it is. As for the contributors, is not not enough that they are listed in the Contributors tab on the repo homepage? I would guess this list will change fairly often, so I think it would be wasted effort to maintain my own list somewhere such as in the README would it not?

@mdbloice - in [this review](https://github.com/openjournals/joss-reviews/issues/348#issuecomment-334753211) we faced a similar issue and the submitting author opened an issue on their repository inviting the other contributors to be authors (most declined). Regardless, it would probably be a good idea to include a short sentence in the paper somewhere acknowledging the other contributors and linking to the GitHub contributors page.
