### 2017-11-02 #432 @mdbloice
Hi @arfon certainly I can do that, and I have now updated the `paper.md` file to contain an acknowledgement and a link to the contributors tab of the repository.
