### 2017-11-02 #432 @sealhuang
@mdbloice Thanks for your work. Given you have described how to test the code, I think it's ok for the compromise on automated tests. So you have addressed all my concerns, and  I'm ready to accept this and recommend it for publication. I think the package would be very helpful for the field. Congrats!
@arfon
