### 2017-11-04 #432 @mdbloice
Hi @arfon thanks for the info, I have done that for version 0.1.5 using Zenodo and as you can see I tried to set the archive using whedon, but was not allowed :)
Anyway, the DOI is above.
Thanks again!
