### 2017-11-06 #432 @arfon
> @arfon - not sure if this is the right place to point this out, but one thing I just noticed is that the link to this review thread is broken on the JOSS page for the paper where it says "View review issue".

Good catch! This should be fixed now.
